Avoid deep-cloning the entire DOM on Parse()
The `parser.ParseDocument(doc)` function deep-clones the DOM within doc to avoid
mutating it during parsing. This operation is expensive, but before it couldn't
be avoided.

This introduces a new function `ParseAndMutate(doc)` which callers can use if
they don't care about doc being potentially mutated. If HTML is coming from an
io.Reader, for example, then it should be fine to mutate the DOM since it's
being discarded anyway.

The following entry points are now using ParseAndMutate:

- parser.Parse()
- FromReader()
- FromURL()